floats to int

+++ src/ParticleSystem.h
@@ -15,10 +15,10 @@ class FramebufferObject;
 class ParticleSystem
 {
 public:
-    ParticleSystem(int texture_width, int texture_height, int canvas_width, int canvas_height, GLfloat scale_p, GLfloat scale_v, GLfloat particle_size, QRgb particle_color);
+    ParticleSystem(int texture_width, int texture_height, int canvas_width, int canvas_height, GLuint scale_p, GLuint scale_v, GLuint particle_size, QRgb particle_color);
 
     /** Encodes a value given the scale */
-    glm::vec2 encode(GLfloat value, GLfloat scale);
+    glm::vec2 encode(GLuint value, GLuint scale);
 
     /** This creates a texture */
     void createTexture(GLuint &textureID);
@@ -54,10 +54,10 @@ private:
     GLuint m_canvas_width;
     GLuint m_canvas_height;
 
-    GLfloat m_scale_p;
-    GLfloat m_scale_v;
+    GLuint m_scale_p;
+    GLuint m_scale_v;
 
-    GLfloat m_particle_size;
+    GLuint m_particle_size;
     QRgb m_particle_color;
 };
